pyproject.toml
[tool.poetry]
name = "fastapi-arangodb"
version = "0.1.0"
description = "Trying fastapi with arangodb"
authors = ["Christopher Pieper <c.pieper@going-public.edu>"]

packages = [
    {include="app"}
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
python-arango = "^7.4.1"
loguru = "^0.6.0"
orjson = "^3.7.8"
pydantic = {extras = ["email"], version = "^1.9.1"}
uvicorn = "^0.18.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]
black = "^22.6.0"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
